Move helper functions to src/ltc/math.rs
